+++ src/providers/Auth/index.ts
@@ -4,3 +4,4 @@ export * from "./useAuth";
 export * from "./authContext";
 export * from "./useAuthContext";
 export * from "./useOptionalAuthContext";
+export * from "./useAuthSessionState";