+++ src/views/index.ts
@@ -1,2 +1,3 @@
 export * from "./NotFoundView";
 export * from "./FeatureUnavailableView";
+export * from "./LegalPage";